fix map

+++ web/js/app_jquery.vmap_1.js
@@ -277,9 +277,9 @@
 
         applyTransformParams: function (scale, transX, transY)
         {
-            var scale = parseInt(scale) || 0;
-            var transX = parseInt(transX) || 0;
-            var transY = parseInt(transY) || 0;
+            var scale = parseFloat(scale) || 0;
+            var transX = parseFloat(transX) || 0;
+            var transY = parseFloat(transY) || 0;
             if (this.mode == 'svg')
             {
                 this.rootGroup.setAttribute('transform', 'scale(' + scale + ') translate(' + transX + ', ' + transY + ')');
